+++ Sources/Packages/Sources/SecureEnclaveSecretKit/SecureEnclaveSecret.swift
@@ -14,29 +14,10 @@ extension SecureEnclave {
         init(
             id: Data,
             name: String,
-            authenticationRequirement: AuthenticationRequirement,
-            publicKey: Data,
-        ) {
-            self.id = id
-            self.name = name
-            self.publicKey = publicKey
-            self.attributes = Attributes(
-                keyType: .init(
-                    algorithm: .ecdsa,
-                    size: 256
-                ),
-                authentication: authenticationRequirement,
-                publicKeyAttribution: nil
-            )
-        }
-
-        init(
-            id: String,
-            name: String,
             publicKey: Data,
             attributes: Attributes
         ) {
-            self.id = Data(id.utf8)
+            self.id = id
             self.name = name
             self.publicKey = publicKey
             self.attributes = attributes
